+++ src/KingmakerGunslinger/Blueprints/GunslingerClassBlueprints.cs
@@ -29,7 +29,8 @@ namespace KingmakerGunslinger.Blueprints
             TargetingLegsBlueprintSet targetingLegs,
             BleedingWoundBlueprintSet bleedingWound,
             ExpertLoadingBlueprintSet expertLoading,
-            LightningReloadBlueprintSet lightningReload)
+            LightningReloadBlueprintSet lightningReload,
+            EvasiveBlueprintSet evasive)
         {
             CharacterClass = characterClass ?? throw new ArgumentNullException("characterClass");
             Progression = progression ?? throw new ArgumentNullException("progression");
@@ -58,6 +59,7 @@ namespace KingmakerGunslinger.Blueprints
                 throw new ArgumentNullException("expertLoading");
             LightningReload = lightningReload ??
                 throw new ArgumentNullException("lightningReload");
+            Evasive = evasive ?? throw new ArgumentNullException("evasive");
         }
         internal BlueprintCharacterClass CharacterClass { get; private set; }
         internal BlueprintProgression Progression { get; private set; }
@@ -79,7 +81,8 @@ namespace KingmakerGunslinger.Blueprints
         internal BleedingWoundBlueprintSet BleedingWound { get; private set; }
         internal ExpertLoadingBlueprintSet ExpertLoading { get; private set; }
         internal LightningReloadBlueprintSet LightningReload { get; private set; }
-        internal int Count { get { return 4 + Grit.Count + Deadeye.Count + Dodge.Count + QuickClear.Count + Nimble.Count + PistolWhip.Count + UtilityShot.Count + GunTraining.Count + DeadShot.Count + StartlingShot.Count + TargetingHead.Count + TargetingTorso.Count + TargetingLegs.Count + BleedingWound.Count + ExpertLoading.Count + LightningReload.Count; } }
+        internal EvasiveBlueprintSet Evasive { get; private set; }
+        internal int Count { get { return 4 + Grit.Count + Deadeye.Count + Dodge.Count + QuickClear.Count + Nimble.Count + PistolWhip.Count + UtilityShot.Count + GunTraining.Count + DeadShot.Count + StartlingShot.Count + TargetingHead.Count + TargetingTorso.Count + TargetingLegs.Count + BleedingWound.Count + ExpertLoading.Count + LightningReload.Count + Evasive.Count; } }
     }
 
     internal sealed class GunslingerClassCatalogPublication
@@ -185,6 +188,8 @@ namespace KingmakerGunslinger.Blueprints
                 ExpertLoadingBlueprints.Register(registry);
             LightningReloadBlueprintSet lightningReload =
                 LightningReloadBlueprints.Register(registry, blackPowder, leadBall);
+            EvasiveBlueprintSet evasive = EvasiveBlueprints.Register(library,
+                registry, grit.Resource, characterClass);
             BlueprintProgression progression = registry.Register<BlueprintProgression>(
                 ProgressionSymbol, () => CreateProgression());
 
@@ -207,6 +212,7 @@ namespace KingmakerGunslinger.Blueprints
             progression.LevelEntries[10].Features.Add(bleedingWound.Feature);
             progression.LevelEntries[10].Features.Add(expertLoading.Feature);
             progression.LevelEntries[10].Features.Add(lightningReload.Feature);
+            progression.LevelEntries[14].Features.Add(evasive.Feature);
             Validate(characterClass, progression, proficiencies, fullBab, goodSave,
                 poorSave, startingPistol, blackPowder, leadBall,
                 simple, martial, lightArmor, firearmProficiency);
@@ -214,7 +220,7 @@ namespace KingmakerGunslinger.Blueprints
                 proficiencies, grit, deadeye, dodge, quickClear, nimble, initiative,
                 pistolWhip, utilityShot, gunTraining, deadShot, startlingShot,
                 targetingHead, targetingTorso, targetingLegs, bleedingWound,
-                expertLoading, lightningReload);
+                expertLoading, lightningReload, evasive);
         }
 
         internal static GunslingerClassCatalogPublication Publish(